Create Universe Packages JSON

Prepares a packages.json file required for R-universe for creating our personal CRAN-like repository on R-universe

Usage

create_universe_pkgs_json(
  group = "KWB-R",
  ignore_pkgs = NULL,
  non_r_packages = kwb.pkgstatus::get_non_r_packages(),
  github_token = Sys.getenv("GITHUB_PAT")
)

Arguments

group

default: "KWB-R"

ignore_pkgs

vector of pkg names that should be ignored to prevent problems with codemeta generation (e.g. "kwb.read" fails due to missing metadata on the license!)

non_r_packages

vector with repos that are not R packages (default: kwb.pkgstatus::get_non_r_packages())

github_token

optionally a Github token in order to access private repositories (default: Sys.getenv("GITHUB_PAT"))

Value

JSON structure as required for packages.json

References

https://ropensci.org/blog/2021/06/22/setup-runiverse/#the-packagesjson-registry-file

Examples

create_universe_pkgs_json()
#> [
#>   {
#>     "package": "algoliar",
#>     "url": "https://github.com/KWB-R/algoliar"
#>   },
#>   {
#>     "package": "aquanes.report",
#>     "url": "https://github.com/KWB-R/aquanes.report"
#>   },
#>   {
#>     "package": "dwc.ar4gw",
#>     "url": "https://github.com/KWB-R/dwc.ar4gw"
#>   },
#>   {
#>     "package": "dwc.wells",
#>     "url": "https://github.com/KWB-R/dwc.wells"
#>   },
#>   {
#>     "package": "fakin.path.app",
#>     "url": "https://github.com/KWB-R/fakin.path.app"
#>   },
#>   {
#>     "package": "fhpredict",
#>     "url": "https://github.com/KWB-R/fhpredict"
#>   },
#>   {
#>     "package": "flextreat.hydrus1d",
#>     "url": "https://github.com/KWB-R/flextreat.hydrus1d"
#>   },
#>   {
#>     "package": "geosalz.phreeqc",
#>     "url": "https://github.com/KWB-R/geosalz.phreeqc"
#>   },
#>   {
#>     "package": "ibathwater",
#>     "url": "https://github.com/KWB-R/ibathwater"
#>   },
#>   {
#>     "package": "keys.lid",
#>     "url": "https://github.com/KWB-R/keys.lid"
#>   },
#>   {
#>     "package": "kwb.abimo",
#>     "url": "https://github.com/KWB-R/kwb.abimo"
#>   },
#>   {
#>     "package": "kwb.barplot",
#>     "url": "https://github.com/KWB-R/kwb.barplot"
#>   },
#>   {
#>     "package": "kwb.base",
#>     "url": "https://github.com/KWB-R/kwb.base"
#>   },
#>   {
#>     "package": "kwb.budget",
#>     "url": "https://github.com/KWB-R/kwb.budget"
#>   },
#>   {
#>     "package": "kwb.code",
#>     "url": "https://github.com/KWB-R/kwb.code"
#>   },
#>   {
#>     "package": "kwb.context",
#>     "url": "https://github.com/KWB-R/kwb.context"
#>   },
#>   {
#>     "package": "kwb.datetime",
#>     "url": "https://github.com/KWB-R/kwb.datetime"
#>   },
#>   {
#>     "package": "kwb.db",
#>     "url": "https://github.com/KWB-R/kwb.db"
#>   },
#>   {
#>     "package": "kwb.default",
#>     "url": "https://github.com/KWB-R/kwb.default"
#>   },
#>   {
#>     "package": "kwb.demeau",
#>     "url": "https://github.com/KWB-R/kwb.demeau"
#>   },
#>   {
#>     "package": "kwb.dswt",
#>     "url": "https://github.com/KWB-R/kwb.dswt"
#>   },
#>   {
#>     "package": "kwb.dwa.m150",
#>     "url": "https://github.com/KWB-R/kwb.dwa.m150"
#>   },
#>   {
#>     "package": "kwb.dwd",
#>     "url": "https://github.com/KWB-R/kwb.dwd"
#>   },
#>   {
#>     "package": "kwb.dygraph",
#>     "url": "https://github.com/KWB-R/kwb.dygraph"
#>   },
#>   {
#>     "package": "kwb.en13508.2",
#>     "url": "https://github.com/KWB-R/kwb.en13508.2"
#>   },
#>   {
#>     "package": "kwb.endnote",
#>     "url": "https://github.com/KWB-R/kwb.endnote"
#>   },
#>   {
#>     "package": "kwb.epanet",
#>     "url": "https://github.com/KWB-R/kwb.epanet"
#>   },
#>   {
#>     "package": "kwb.event",
#>     "url": "https://github.com/KWB-R/kwb.event"
#>   },
#>   {
#>     "package": "kwb.fakin",
#>     "url": "https://github.com/KWB-R/kwb.fakin"
#>   },
#>   {
#>     "package": "kwb.fcr",
#>     "url": "https://github.com/KWB-R/kwb.fcr"
#>   },
#>   {
#>     "package": "kwb.file",
#>     "url": "https://github.com/KWB-R/kwb.file"
#>   },
#>   {
#>     "package": "kwb.flusshygiene",
#>     "url": "https://github.com/KWB-R/kwb.flusshygiene"
#>   },
#>   {
#>     "package": "kwb.flusshygiene.app",
#>     "url": "https://github.com/KWB-R/kwb.flusshygiene.app"
#>   },
#>   {
#>     "package": "kwb.geosalz",
#>     "url": "https://github.com/KWB-R/kwb.geosalz"
#>   },
#>   {
#>     "package": "kwb.gocr",
#>     "url": "https://github.com/KWB-R/kwb.gocr"
#>   },
#>   {
#>     "package": "kwb.graph",
#>     "url": "https://github.com/KWB-R/kwb.graph"
#>   },
#>   {
#>     "package": "kwb.hantush",
#>     "url": "https://github.com/KWB-R/kwb.hantush"
#>   },
#>   {
#>     "package": "kwb.heatsine",
#>     "url": "https://github.com/KWB-R/kwb.heatsine"
#>   },
#>   {
#>     "package": "kwb.heatsine.opencpu",
#>     "url": "https://github.com/KWB-R/kwb.heatsine.opencpu"
#>   },
#>   {
#>     "package": "kwb.hydrus1d",
#>     "url": "https://github.com/KWB-R/kwb.hydrus1d"
#>   },
#>   {
#>     "package": "kwb.impetus",
#>     "url": "https://github.com/KWB-R/kwb.impetus"
#>   },
#>   {
#>     "package": "kwb.infoworks",
#>     "url": "https://github.com/KWB-R/kwb.infoworks"
#>   },
#>   {
#>     "package": "kwb.iview",
#>     "url": "https://github.com/KWB-R/kwb.iview"
#>   },
#>   {
#>     "package": "kwb.kuras",
#>     "url": "https://github.com/KWB-R/kwb.kuras"
#>   },
#>   {
#>     "package": "kwb.kuras.db",
#>     "url": "https://github.com/KWB-R/kwb.kuras.db"
#>   },
#>   {
#>     "package": "kwb.lca",
#>     "url": "https://github.com/KWB-R/kwb.lca"
#>   },
#>   {
#>     "package": "kwb.logger",
#>     "url": "https://github.com/KWB-R/kwb.logger"
#>   },
#>   {
#>     "package": "kwb.matomo",
#>     "url": "https://github.com/KWB-R/kwb.matomo"
#>   },
#>   {
#>     "package": "kwb.mia.evalCritO2",
#>     "url": "https://github.com/KWB-R/kwb.mia.evalCritO2"
#>   },
#>   {
#>     "package": "kwb.mia.iw",
#>     "url": "https://github.com/KWB-R/kwb.mia.iw"
#>   },
#>   {
#>     "package": "kwb.miacso",
#>     "url": "https://github.com/KWB-R/kwb.miacso"
#>   },
#>   {
#>     "package": "kwb.misa",
#>     "url": "https://github.com/KWB-R/kwb.misa"
#>   },
#>   {
#>     "package": "kwb.misc",
#>     "url": "https://github.com/KWB-R/kwb.misc"
#>   },
#>   {
#>     "package": "kwb.ml",
#>     "url": "https://github.com/KWB-R/kwb.ml"
#>   },
#>   {
#>     "package": "kwb.monitoring",
#>     "url": "https://github.com/KWB-R/kwb.monitoring"
#>   },
#>   {
#>     "package": "kwb.nextcloud",
#>     "url": "https://github.com/KWB-R/kwb.nextcloud"
#>   },
#>   {
#>     "package": "kwb.odm",
#>     "url": "https://github.com/KWB-R/kwb.odm"
#>   },
#>   {
#>     "package": "kwb.odmx",
#>     "url": "https://github.com/KWB-R/kwb.odmx"
#>   },
#>   {
#>     "package": "kwb.ogre",
#>     "url": "https://github.com/KWB-R/kwb.ogre"
#>   },
#>   {
#>     "package": "kwb.ogre.model",
#>     "url": "https://github.com/KWB-R/kwb.ogre.model"
#>   },
#>   {
#>     "package": "kwb.orcid",
#>     "url": "https://github.com/KWB-R/kwb.orcid"
#>   },
#>   {
#>     "package": "kwb.package",
#>     "url": "https://github.com/KWB-R/kwb.package"
#>   },
#>   {
#>     "package": "kwb.pathdict",
#>     "url": "https://github.com/KWB-R/kwb.pathdict"
#>   },
#>   {
#>     "package": "kwb.pilot",
#>     "url": "https://github.com/KWB-R/kwb.pilot"
#>   },
#>   {
#>     "package": "kwb.pkgbuild",
#>     "url": "https://github.com/KWB-R/kwb.pkgbuild"
#>   },
#>   {
#>     "package": "kwb.pkgstatus",
#>     "url": "https://github.com/KWB-R/kwb.pkgstatus"
#>   },
#>   {
#>     "package": "kwb.plot",
#>     "url": "https://github.com/KWB-R/kwb.plot"
#>   },
#>   {
#>     "package": "kwb.prep",
#>     "url": "https://github.com/KWB-R/kwb.prep"
#>   },
#>   {
#>     "package": "kwb.pubs",
#>     "url": "https://github.com/KWB-R/kwb.pubs"
#>   },
#>   {
#>     "package": "kwb.python",
#>     "url": "https://github.com/KWB-R/kwb.python"
#>   },
#>   {
#>     "package": "kwb.qmra",
#>     "url": "https://github.com/KWB-R/kwb.qmra"
#>   },
#>   {
#>     "package": "kwb.quantum",
#>     "url": "https://github.com/KWB-R/kwb.quantum"
#>   },
#>   {
#>     "package": "kwb.rain",
#>     "url": "https://github.com/KWB-R/kwb.rain"
#>   },
#>   {
#>     "package": "kwb.read",
#>     "url": "https://github.com/KWB-R/kwb.read"
#>   },
#>   {
#>     "package": "kwb.readxl",
#>     "url": "https://github.com/KWB-R/kwb.readxl"
#>   },
#>   {
#>     "package": "kwb.resilience",
#>     "url": "https://github.com/KWB-R/kwb.resilience"
#>   },
#>   {
#>     "package": "kwb.satellite",
#>     "url": "https://github.com/KWB-R/kwb.satellite"
#>   },
#>   {
#>     "package": "kwb.site",
#>     "url": "https://github.com/KWB-R/kwb.site"
#>   },
#>   {
#>     "package": "kwb.svn",
#>     "url": "https://github.com/KWB-R/kwb.svn"
#>   },
#>   {
#>     "package": "kwb.swmm",
#>     "url": "https://github.com/KWB-R/kwb.swmm"
#>   },
#>   {
#>     "package": "kwb.test",
#>     "url": "https://github.com/KWB-R/kwb.test"
#>   },
#>   {
#>     "package": "kwb.twitter",
#>     "url": "https://github.com/KWB-R/kwb.twitter"
#>   },
#>   {
#>     "package": "kwb.ucode",
#>     "url": "https://github.com/KWB-R/kwb.ucode"
#>   },
#>   {
#>     "package": "kwb.umberto",
#>     "url": "https://github.com/KWB-R/kwb.umberto"
#>   },
#>   {
#>     "package": "kwb.utils",
#>     "url": "https://github.com/KWB-R/kwb.utils"
#>   },
#>   {
#>     "package": "kwb.virtualdub",
#>     "url": "https://github.com/KWB-R/kwb.virtualdub"
#>   },
#>   {
#>     "package": "kwb.vs2dh",
#>     "url": "https://github.com/KWB-R/kwb.vs2dh"
#>   },
#>   {
#>     "package": "kwb.wtaq",
#>     "url": "https://github.com/KWB-R/kwb.wtaq"
#>   },
#>   {
#>     "package": "pkgmeta",
#>     "url": "https://github.com/KWB-R/pkgmeta"
#>   },
#>   {
#>     "package": "qmra.db",
#>     "url": "https://github.com/KWB-R/qmra.db"
#>   },
#>   {
#>     "package": "r2q",
#>     "url": "https://github.com/KWB-R/r2q"
#>   },
#>   {
#>     "package": "sema.berlin.utils",
#>     "url": "https://github.com/KWB-R/sema.berlin.utils"
#>   },
#>   {
#>     "package": "urbanAnnualRunoff",
#>     "url": "https://github.com/KWB-R/urbanAnnualRunoff"
#>   },
#>   {
#>     "package": "wasserportal",
#>     "url": "https://github.com/KWB-R/wasserportal"
#>   }
#> ]